+++ frontend/src/queries/category/useCategoryListQuery.spec.tsx
@@ -1,8 +1,8 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 
+import { AuthProvider } from '@/contexts';
 import { categories as mockCategories } from '@/mocks/categoryList.json';
-import { AuthProvider } from '../../contexts/authContext';
 import { useCategoryListQuery } from './useCategoryListQuery';
 
 const queryClient = new QueryClient();
