+++ frontend/src/queries/tag/useTagListQuery.spec.tsx
@@ -1,8 +1,8 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 
+import { AuthProvider } from '@/contexts';
 import { tags as mockTags } from '@/mocks/tagList.json';
-import { AuthProvider } from '../../contexts/authContext';
 import { useTagListQuery } from '../tag/useTagListQuery';
 
 const queryClient = new QueryClient();
