+++ frontend/src/queries/template/useTemplateListQuery.spec.tsx
@@ -1,7 +1,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { renderHook, waitFor } from '@testing-library/react';
 
-import { AuthProvider } from '@/contexts/authContext';
+import { AuthProvider } from '@/contexts';
 import { templates as mockTemplates } from '@/mocks/templateList.json';
 import { DEFAULT_SORTING_OPTION, PAGE_SIZE } from '../../api/templates';
 import { useTemplateListQuery } from './useTemplateListQuery';
